fix: 구매저/참가자 관리 조회 주문 상태 응답 수정

+++ src/main/java/com/gotogether/domain/hostchannel/service/HostChannelServiceImpl.java
@@ -193,7 +193,7 @@ public class HostChannelServiceImpl implements HostChannelService {
 			List<Order> orders = orderRepository.findByTicketAndStatus(ticket, OrderStatus.COMPLETED);
 
 			totalTicketCnt += orders.size();
-			totalPrice += orders.size() * ticket.getPrice();
+			totalPrice += (long)orders.size() * ticket.getPrice();
 		}
 
 		return HostChannelConverter.toHostDashboardResponseDTO(event, totalTicketCnt, totalPrice);
